// app/src/components/CodeEditor/types.ts
export interface ThemeColors {
  // Main colors
  bg: string
  fg: string
  accent: string
  
  // UI component colors
  menuBarBg: string
  menuBarFg: string
  menuBarBorder: string
  menuDropdownBg: string
  menuDropdownFg: string
  menuDropdownHover: string
  menuDropdownBorder: string
  
  headerBg: string
  headerFg: string
  headerBorder: string
  
  sidebarBg: string
  sidebarFg: string
  sidebarBorder: string
  sidebarItemHover: string
  sidebarItemSelected: string
  
  editorBg: string
  editorFg: string
  
  aiPanelBg: string
  aiPanelFg: string
  aiPanelBorder: string
  aiMessageBg: string
  aiInputBg: string
  aiInputFg: string
  aiInputBorder: string
  
  terminalBg: string
  terminalFg: string
  terminalBorder: string
  
  dialogBg: string
  dialogFg: string
  dialogBorder: string
  dialogHeaderBg: string
  
  buttonBg: string
  buttonFg: string
  buttonHover: string
  buttonPrimaryBg: string
  buttonPrimaryFg: string
  
  inputBg: string
  inputFg: string
  inputBorder: string
  inputFocusBorder: string
  
  tabBg: string
  tabFg: string
  tabActiveBg: string
  tabActiveFg: string
  tabBorder: string
  
  statusBarBg: string
  statusBarFg: string
  statusBarBorder: string
}